requirements.txt
pymongo==3.7.2
python-decouple==3.1
responses==0.10.5
